### 2025-07-24 #8639 @editorialbot
[PRE REVIEW]: ReciPies: A Lightweight Data Transformation Pipeline for Reproducible ML
**Submitting author:** @rvandewater (<a href="http://orcid.org/0000-0002-2895-4872">Robin van de Water</a>)
**Repository:** https://github.com/rvandewater/ReciPies
**Branch with paper.md** (empty if default branch):
**Version:** v1.2.0
