### 2025-07-25 936c520
remove character
