### 2025-07-25 #8639 @rvandewater
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/16505494510?check_suite_focus=true). Problem with ORCID (0000-0002-2895-4872ß) for Robin van de Water. ORCID looks to be the wrong length.

Removed character
