### 2025-07-25 #8639 @crvernon
@rvandewater while I am getting you setup with a topic editor, please reduce the word count of your paper to <= 1000 words.  We are working through a large backlog of submissions and will get to yours as soon as possible.
