### 2025-07-25 fac3a7e
reduce paper length and fixed some formatting
