+++ paper/paper.md
@@ -1,19 +1,19 @@
 ---
 title: >-
-    `ReciPies`: A Lightweight Data Transformation Pipeline for Reproducible ML
+  `ReciPies`: A Lightweight Data Transformation Pipeline for Reproducible ML
 authors:
   - name: Robin P. van de Water
     email: robin.vandewater@hpi.de
-    affiliation: [1, 2]
+    affiliation: [ 1, 2 ]
     orcid: 0000-0002-2895-4872
     corresponding: true
   - name: Hendrik Schmidt
     orcid: 0000-0001-7699-3983
-    affiliation: 
+    affiliation:
     equal-contrib: false
   - name: Patrick Rockenschaub
     orcid: 0000-0002-6499-7933
-    affiliation: [1, 3]
+    affiliation: [ 1, 3 ]
     equal-contrib: false
 affiliations:
   - index: 1
@@ -31,54 +31,90 @@ tags:
   - markdown
   - publishing
 ---
-<!-- 
-Guide:
-https://github.com/openjournals/inara/blob/main/example/paper.md  -->
-<!-- # Summary
-
-Machine learning pipelines often require complicated preprocessing pipelines. Our aim is to simplify this with a python package that can be used to define human-readable and reproducible pipelines for machine learning tasks.
-
-# Statement of Need
-
-Python is the most popular programming language for machine learning. However, preprocessing data for machine learning tasks can be a time-consuming and error-prone process. `ReciPies` is a python package that aims to simplify this process by providing a fast and intuitive interface for preprocessing data. It can use both a polars [@PolarsPolars2024] backend, which allows for fast python-native data processing, and a traditional Pandas [@mckinney-proc-scipy-2010] backend for use with legacy data tools. 
-`ReciPies` is designed to be easy to use and flexible, allowing users to easily preprocess data for a wide range of machine learning pipelines. 
-Moreover, we hide the complexity of the preprocessing pipeline from the user which allows for more reproducible and maintainable code. -->
 
 # Summary
-Modern machine learning (ML) workflows live or die by their data‑preprocessing steps, yet in Python—a language with a rich ecosystem for data science and ML—these steps are often scattered across ad‑hoc scripts or opaque Scikit-Learn (sklearn) snippets that are hard to read, audit, or reuse. `ReciPies` provides a concise, human‑readable, and fully reproducible way to declare, execute, and share preprocessing pipelines, adhering to Configuration as Code principles. It lets users describe transformations as a recipe made of ordered *steps* (e.g., imputing, encoding, normalizing) applied to variables identified by semantic roles (predictor, outcome, ID, time stamp, etc.). Recipes can be *prepped* (trained) once, *baked* many times, and cleanly separated between training and new data—preventing data leakage by construction. Under the hood, `ReciPies` targets both Pandas and Polars backends for performance and flexibility, and it is easily extensible: users can register custom steps with minimal boilerplate. Each recipe is serializable to JSON/YAML for provenance tracking, collaboration, and publication, and integrates smoothly with downstream modeling libraries. Packaging preprocessing as clear, declarative objects, `ReciPies` lowers the cognitive load of feature engineering, improves reproducibility, and makes methodological choices explicit, benefiting individual researchers, engineering teams, and peer reviewers alike.
+
+Machine Learning (ML) workflows live or die by their data‑preprocessing steps, yet in Python, these steps are often
+scattered across ad‑hoc scripts or opaque Scikit-Learn (sklearn) snippets that are hard to read, audit, or reuse.
+`ReciPies` provides a concise, human‑readable, and fully
+reproducible way to declare, execute, and share preprocessing pipelines, that adheres to Configuration as Code
+principles.
+It lets users describe transformations as a recipe made of ordered *steps* (e.g., imputing, encoding, normalizing)
[…]
