+++ paper/paper.md
@@ -9,7 +9,7 @@ authors:
     corresponding: true
   - name: Hendrik Schmidt
     orcid: 0000-0001-7699-3983
-    affiliation:
+    affiliation: [ 1 ]
     equal-contrib: false
   - name: Patrick Rockenschaub
     orcid: 0000-0002-6499-7933
@@ -19,17 +19,13 @@ affiliations:
   - index: 1
     name: Hasso Plattner Institute, University of Potsdam, Potsdam, Germany
   - index: 2
-    name: Hasso Plattner Institute for Digital Health at Mount Sinai, Icahn School of Medicine at Mount Sinai, New York City, NY, USA
+    name: Icahn School of Medicine at Mount Sinai, New York City, NY, USA
   - index: 3
     name: Innsbruck Medical University, Innsbruck, Austria
-date: 2025-07-24
+date: 2025-07-25
 bibliography: paper.bib
-repository: https://github.com/rvandewater/`ReciPies`
+repository: https://github.com/rvandewater/ReciPies
 tags:
-  - reference
-  - example
-  - markdown
-  - publishing
 ---
 
 # Summary
@@ -41,9 +37,9 @@ reproducible way to declare, execute, and share preprocessing pipelines, that ad
 principles.
 It lets users describe transformations as a recipe made of ordered *steps* (e.g., imputing, encoding, normalizing)
 applied to variables identified by semantic roles (predictor, outcome, ID, time stamp, etc.). Recipes can be *prepped*
-once, *baked* many times, and cleanly separated between training and new data. `ReciPies` provides the choice of both
+once, *baked* many times, and separated between training and new data. `ReciPies` provides the choice of
 Pandas and Polars backends and
-is easily extensible: users can create steps with minimal boilerplate. Provenance can be tracked and published.
+is easily extensible. Data provenance can be tracked and published.
 Packaging preprocessing as clear, declarative objects, `ReciPies` lowers the cognitive load of feature engineering,
 improves reproducibility, and makes methodological choices explicit, benefiting individual researchers, engineering
 teams, and peer reviewers alike.
@@ -52,9 +48,8 @@ teams, and peer reviewers alike.
 
 Robust machine‑learning results in science hinge on transparent, reproducible data‑preprocessing—yet in Python, these
 steps are typically spread across ad‑hoc notebooks or are buried inside opaque scripts; that is, if this code is even
-made available. Additionally, most variable semantics are unclear (*Which columns are outcomes? IDs? Time stamps?*);
-this encourages accidental data leakage when "fitting" transforms on the full dataset, confounding research results,
-complicating peer review, and hindering reuse. Researchers and engineers working with longitudinal or regulated data (
+made available. Additionally, most variable semantics are unclear. These problems confound research results,
+complicate peer review, and hinder reuse. Researchers and engineers working with longitudinal regulated data (
 e.g., energy production, finance, and environmental monitoring) especially need pipelines they are able to audit,
 serialize, and hand to collaborators without reverse‑engineering a tangle of imperative code [@10.1145/3641525]. The
 lack of reproducibility has been documented extensively in
@@ -70,26 +65,19 @@ Object-oriented abstractions enable users to implement custom steps.
 
 Our work is an improvement upon [@kuhnRecipesPreprocessingFeature2024]  aimed at the ML
 community. The design enables
-straightforward integration as part of a pipeline that includes machine learning libraries such as
+straightforward integration as part of a pipeline that includes machine learning libraries like
 sklearn [@pedregosa_scikit-learn_2011] and PyTorch[@paszkePyTorchImperativeStyle2019]. To the best of our knowledge, no
[…]
