### 2025-10-15 #9261 @editorialbot
[REVIEW]: ReciPies: A Lightweight Data Transformation Pipeline for Reproducible ML
**Submitting author:** @rvandewater (<a href="http://orcid.org/0000-0002-2895-4872">Robin van de Water</a>)
**Repository:** https://github.com/rvandewater/ReciPies
**Branch with paper.md** (empty if default branch):
**Version:** 1.3.1
