### 2025-10-15 #9261 @crvernon
👋 @rvandewater , @simonprovost , and @panagiotisanagnostou - This is the review thread for the paper. All of our communications will happen here from now on.
