+++ paper/paper.md
@@ -4,22 +4,22 @@ title: >-
 authors:
   - name: Robin P. van de Water
     email: robin.vandewater@hpi.de
-    affiliation: [ 1, 2 ]
+    affiliation: "1, 2"
     orcid: 0000-0002-2895-4872
     corresponding: true
   - name: Hendrik Schmidt
     orcid: 0000-0001-7699-3983
-    affiliation: [ 1 ]
+    affiliation: "1"
     equal-contrib: false
   - name: Patrick Rockenschaub
     orcid: 0000-0002-6499-7933
-    affiliation: [ 1, 3 ]
+    affiliation: "3"
     equal-contrib: false
 affiliations:
   - index: 1
     name: Hasso Plattner Institute, University of Potsdam, Potsdam, Germany
   - index: 2
-    name: Icahn School of Medicine at Mount Sinai, New York City, NY, USA
+    name: Hasso Plattner Institute for Digital Health at Mount Sinai, Icahn School of Medicine at Mount Sinai, New York City, NY, USA
   - index: 3
     name: Innsbruck Medical University, Innsbruck, Austria
 date: 2025-07-25
