+++ paper/paper.bib
@@ -1,8 +1,11 @@
-@book{10.1145/3641525,
+@inproceedings{10.1145/3641525,
   title = {{{ACM REP}} '24: {{Proceedings}} of the 2nd {{ACM}} Conference on Reproducibility and Replicability},
+  booktitle = {{{ACM REP}} '24: {{Proceedings}} of the 2nd {{ACM}} Conference on Reproducibility and Replicability},
+  author = {Various, Authors},
   date = {2024},
   publisher = {Association for Computing Machinery},
   location = {Rennes, France and New York, NY, USA},
+  doi = {10.1145/3641525},
   isbn = {979-8-4007-0530-4}
 }
 
@@ -77,6 +80,7 @@
   abstract = {Mortality prediction of intensive care unit (ICU) patients facilitates hospital benchmarking and has the opportunity to provide caregivers with useful summaries of patient health at the bedside. The development of novel models for mortality prediction is a popular task in machine learning, with researchers typically seeking to maximize measures such as the area under the receiver operator characteristic curve (AUROC). The number of ’researcher degrees of freedom’ that contribute to the performance of a model, however, presents a challenge when seeking to compare reported performance of such models. In this study, we review publications that have reported performance of mortality prediction models based on the Medical Information Mart for Intensive Care (MIMIC) database and attempt to reproduce the cohorts used in their studies. We then compare the performance reported in the studies against gradient boosting and logistic regression models using a simple set of features extracted from MIMIC. We demonstrate the large heterogeneity in studies that purport to conduct the single task of ’mortality prediction’, highlighting the need for improvements in the way that prediction tasks are reported to enable fairer comparison between models. We reproduced datasets for 38 experiments corresponding to 28 published studies using MIMIC. In half of the experiments, the sample size we acquired was 25\% greater or smaller than the sample size reported. The highest discrepancy was 11,767 patients. While accurate reproduction of each study cannot be guaranteed, we believe that these results highlight the need for more consistent reporting of model design and methodology to allow performance improvements to be compared. We discuss the challenges in reproducing the cohorts used in the studies, highlighting the importance of clearly reported methods (e.g. data cleansing, variable selection, cohort selection) and the need for open code and publicly available benchmarks.},
   eventtitle = {Machine {{Learning}} for {{Healthcare Conference}}},
   langid = {english},
+  doi = {10.1038/s41597-022-01899-x},
   file = {/Users/robin/Zotero/storage/5C74RYC5/Johnson et al. - 2017 - Reproducibility in critical care a mortality prediction case study.pdf}
 }
 
@@ -106,7 +110,9 @@
   url = {https://cloud.r-project.org/web/packages/recipes/index.html},
   urldate = {2024-07-09},
   abstract = {A recipe prepares your data for modeling. We provide an extensible framework for pipeable sequences of feature engineering steps provides preprocessing tools to be applied to data. Statistical parameters for the steps can be estimated from an initial data set and then applied to other data sets. The resulting processed output can then be used as inputs for statistical or machine learning models.},
-  version = {1.1.0}
+  version = {1.1.0},
+  annotation = {10.32614/CRAN.package.recipes},
+  doi = {10.32614/CRAN.package.recipes},
 }
 
 @inproceedings{mckinney-proc-scipy-2010,
@@ -127,6 +133,7 @@
   eprinttype = {arXiv},
   eprintclass = {cs},
   url = {http://arxiv.org/abs/2312.11122},
+  doi = {10.48786/edbt.2025.27}
   urldate = {2024-07-09},
[…]
