+++ paper/paper.bib
@@ -133,7 +133,7 @@
   eprinttype = {arXiv},
   eprintclass = {cs},
   url = {http://arxiv.org/abs/2312.11122},
-  doi = {10.48786/edbt.2025.27}
+  doi = {10.48786/edbt.2025.27},
   urldate = {2024-07-09},
   abstract = {Data preparation is a trial-and-error process that typically involves countless iterations over the data to define the best pipeline of operators for a given task. With tabular data, practitioners often perform that burdensome activity on local machines by writing ad hoc scripts with libraries based on the Pandas dataframe API and testing them on samples of the entire dataset—the faster the library, the less idle time its users have.},
   langid = {english},
@@ -169,7 +169,7 @@
   issn = {1533-7928},
   url = {http://jmlr.org/papers/v12/pedregosa11a.html},
   urldate = {2022-12-21},
-  doi = {10.48550/arXiv.1201.0490}
+  doi = {10.48550/arXiv.1201.0490},
   abstract = {Scikit-learn is a Python module integrating a wide range of state-of-the-art machine learning algorithms for medium-scale supervised and unsupervised problems. This package focuses on bringing machine learning to non-specialists using a general-purpose high-level language. Emphasis is put on ease of use, performance, documentation, and API consistency. It has minimal dependencies and is distributed under the simplified BSD license, encouraging its use in both academic and commercial settings. Source code, binaries, and documentation can be downloaded from http://scikit-learn.sourceforge.net.}
 }
 
