### 2025-11-01 rvandewater/recipys#24 @simonprovost
[JOSS-Review] Code of Conduct Typos
Hi folks!

May you update the Code of Conduct to be reflecting your chose point of contact please?

```ref
Instances of abusive, harassing, or otherwise unacceptable behavior may be reported to the project maintainers at [YOUR EMAIL HERE]. All complaints will be reviewed and investigated promptly and fairly.
```

Cheers
