### 2025-11-01 rvandewater/recipys#25 @simonprovost
[JOSS-Review] Readme left chunk ?
Hi folks,

Is the paragraph in the License section of the readme, as follows, supposed to be there?

```
To define preprocessing operations, one has to supply roles to the different columns of the Dataframe. This allows the user to create groups of columns which have a particular function. Then, we provide several "steps" that can be applied to the datasets, among which: Historical accumulation, Resampling the time resolution, A number of imputation methods, and a wrapper for any [Scikit-learn](https://github.com/scikit-learn/scikit-learn) preprocessing step. We believe to have covered any basic preprocessing needs for prepared datasets. Any missing step can be added by following the step interface.
```

To remove if not, please

Cheers
