### 2025-11-01 rvandewater/recipys#27 @simonprovost
[JOSS-Review] Move `icu-demo-data`
Hi folks!

To have a less messy folders-to-files organisation in the repo, it'd not be too bad to think about moving the `icu-demo-data` into data a `data/` generic folder or into `examples/` if only used in `examples` —— Reducing overall confusions

If aga it let me know why please, I do not mind

Cheers
