### 2025-11-01 rvandewater/recipys#28 @simonprovost
[JOSS-Review] May you add to your gitignore `.idea/`
Hi folks,

An easy rapid change which is best-practice especially in OSS (alright in private-team-based repo possibly) is to remove your `.idea/` folder and adding it to your gitignore please.

if aga it let me know why please

```
# JetBrains IDE project files
.idea/
```

Cheers
