### 2025-11-01 rvandewater/recipys#29 @simonprovost
[JOSS-Review] Planning into a `main/` release soon?
Hi folks,

May I know why is `main/` miles away from your curr. `dev` branch? Is this planned to be released + tagged once the JOSS paper is published?  If so you can auto-close that issue, please. Otherwise, please either confirm a soon release onto `main` or state why wouldn't you, please?

Cheers
