### 2025-11-01 rvandewater/recipys#30 @simonprovost
[JOSS-Review] Improve Documentation
Hi folks,

I would appreciate it if the documentation could:

1) Point to your API ref. much easier. Either by horizontal-tabbing at the top, or removing the `contents` appearing on the vertical-sidebar tabbing.

2) Do you really need to let the constructors of your classes in your doc? It appears that the docstrings are class-wise anyway so it adds length for not much? Am I mistaken, if so could you show where it is useful please?

3) It would be much better to have a `tutorials/` section in your documentation, going beyond the given `example/`. For instance, (1) Your first ReciPIes pipeline, (2) How to switch `backends`? and (3) Advanced ReciPIes pipeline. <--- Random thoughts but hope you’ve got the point of some tutorials to help newcomers.

4) May we have interactively the `examples/` notebooks in a tab `Examples` too? Seeing the interactive notebook (see an example here: https://urbanmapper.readthedocs.io/en/latest/copy_of_examples/1-Per-Module/7-urban_pipeline/#running-the-pipeline — If not useful then let it there as is but people look for such stuff. I would also argue that this could be merged with the tutorials somehow. Let me know your thoughts.

5) Think of having a new tab `Concepts` where you will be able to go more in depth with the potential viz. from #26 if ever was chosen to be taken into account, the core concepts, and the backends. How everything is cooking itself together. Newcomers might need such material. You are introducing a new concept, it better be understood however brilliant it is otherwise people will not use it. You mention Sklearn a lot, Sklearn documentation is a goldmine, let's make ReciPies' one too.

Cheers.
