### 2025-11-01 #9261 @simonprovost
# **Simon Peer Review of the repo's content**

See my _per-repo-github_ issues at:

- https://github.com/rvandewater/ReciPies/issues/24
- https://github.com/rvandewater/ReciPies/issues/25
- https://github.com/rvandewater/ReciPies/issues/26
- https://github.com/rvandewater/ReciPies/issues/27
- https://github.com/rvandewater/ReciPies/issues/28
- https://github.com/rvandewater/ReciPies/issues/29
- https://github.com/rvandewater/ReciPies/issues/30

Cheers —— Will complete with my overall thoughts! 🫡
