### 2025-11-03 89947d0
added precommit and fix gitignore
