+++ benchmarking/generate_data.py
@@ -1,16 +1,21 @@
-
 import polars as pl
 import numpy as np
 from datetime import datetime, timedelta
 import random
+
 # Enable StringCache
 pl.StringCache()
 
 
-
-
-def generate_generic_datetime_data(stay_ids: list[int], max_rows: int, start_time: datetime, timestep: timedelta, columns: list[str], missingness: dict[str, float],
-                                   categorical_columns=None) -> pl.DataFrame:
+def generate_generic_datetime_data(
+    stay_ids: list[int],
+    max_rows: int,
+    start_time: datetime,
+    timestep: timedelta,
+    columns: list[str],
+    missingness: dict[str, float],
+    categorical_columns=None,
+) -> pl.DataFrame:
     """
     Generate a Polars DataFrame with the specified number of rows and schema for multiple stay IDs.
 
@@ -26,7 +31,7 @@ def generate_generic_datetime_data(stay_ids: list[int], max_rows: int, start_tim
     - A Polars DataFrame with the generated data.
     """
     # Distribute rows randomly among stay IDs
-    rows_per_stay_id = np.random.multinomial(max_rows, [1/len(stay_ids)]*len(stay_ids))
+    rows_per_stay_id = np.random.multinomial(max_rows, [1 / len(stay_ids)] * len(stay_ids))
 
     all_data = []
 
@@ -35,10 +40,7 @@ def generate_generic_datetime_data(stay_ids: list[int], max_rows: int, start_tim
         time_column = [start_time + i * timestep for i in range(num_rows)]
 
         # Initialize the data dictionary with the time and stay_id columns
-        data = {
-            "stay_id": [stay_id] * num_rows,
-            "time": time_column
-        }
+        data = {"stay_id": [stay_id] * num_rows, "time": time_column}
 
         # Generate random data for other columns
         for column in columns:
@@ -64,7 +66,6 @@ def generate_generic_datetime_data(stay_ids: list[int], max_rows: int, start_tim
                 data[column] = pl.Series(data[column], dtype=pl.Categorical)
                 # data[column] = pl.Series(data[column], dtype=pl.Enum)
 
-
         # Create the DataFrame for the current stay_id
         df = pl.DataFrame(data)
         all_data.append(df)
@@ -74,28 +75,74 @@ def generate_generic_datetime_data(stay_ids: list[int], max_rows: int, start_tim
 
     return final_df
 
-def generate_icu_data(max_rows: int, missingness_threshold: tuple[int,int]=(0.1,0.7), seed:int=None) -> pl.DataFrame:
+
+def generate_icu_data(max_rows: int, missingness_threshold: tuple[int, int] = (0.1, 0.7), seed: int = None) -> pl.DataFrame:
     # Example usage
     if seed is not None:
         random.seed(seed)
-    stay_ids = random.sample(range(1, max_rows*100), int(max_rows/24))
+    stay_ids = random.sample(range(1, max_rows * 100), int(max_rows / 24))
     start_time = datetime(2024, 1, 1)
     timestep = timedelta(hours=1)
 
     # Define the list of column names
-    columns = ["alb", "alp", "alt", "ast", "be", "bicar", "bili", "bili_dir", "bnd", "bun", "ca", "cai", "ck",
-               "ckmb", "cl", "crea", "crp", "dbp", "fgn", "fio2", "glu", "hgb", "hr", "inr_pt", "k", "lact", "lymph", "map",
-               "mch", "mchc", "mcv", "methb", "mg", "na", "neut", "o2sat", "pco2", "ph", "phos", "plt", "po2", "ptt", "resp",
-               "sbp", "temp", "tnt", "urine", "wbc"]
[…]
