+++ docs/source/conf.py
@@ -1,6 +1,7 @@
 import os
 import sys
-sys.path.insert(0, os.path.abspath('../../src'))  # Add your project root to Python path
+
+sys.path.insert(0, os.path.abspath("../../src"))  # Add your project root to Python path
 # Configuration file for the Sphinx documentation builder.
 #
 # For the full list of built-in configuration values, see the documentation:
@@ -9,16 +10,16 @@ sys.path.insert(0, os.path.abspath('../../src'))  # Add your project root to Pyt
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = 'ReciPies'
-copyright = '2025, Robin P. van de Water'
-author = 'Robin P. van de Water'
-release = ''
+project = "ReciPies"
+copyright = "2025, Robin P. van de Water"
+author = "Robin P. van de Water"
+release = ""
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 
-templates_path = ['_templates']
+templates_path = ["_templates"]
 exclude_patterns = []
 
 
@@ -26,7 +27,7 @@ exclude_patterns = []
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
 # Theme configuration
-html_theme = 'sphinx_immaterial'
+html_theme = "sphinx_immaterial"
 # Theme options
 html_theme_options = {
     # Toc options
@@ -34,20 +35,20 @@ html_theme_options = {
 
 
 extensions = [
-    'sphinx.ext.autodoc',
-    'sphinx.ext.napoleon',
-    'sphinx.ext.viewcode',
-    'myst_parser',
-    'sphinx_immaterial',
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.viewcode",
+    "myst_parser",
+    "sphinx_immaterial",
 ]
 
 # Autodoc settings
 autodoc_default_options = {
-    'members': True,
-    'member-order': 'bysource',
-    'special-members': '__init__',
-    'undoc-members': True,
-    'exclude-members': '__weakref__'
+    "members": True,
+    "member-order": "bysource",
+    "special-members": "__init__",
+    "undoc-members": True,
+    "exclude-members": "__weakref__",
 }
 
 # Generate autosummary even if no references
