+++ docs/source/api.rst
@@ -17,4 +17,3 @@ API Reference
    :members:
    :undoc-members:
    :show-inheritance:
-

[…]
