### 2025-11-04 68045f8
readme improvements and flowchart
