### 2025-11-04 905a17e
update readability. created white background to ensure readable in dark mode
