### 2025-11-04 845f08e
readme diagram description
