### 2025-11-04 a82a918
added flowchart to paper
