### 2025-11-04 3cba88a
docs dependencies
