### 2025-11-04 rvandewater/recipys#26 @rvandewater
@simonprovost I have added a diagram to show how recipies works @simonprovost. I do see your other suggestions and I have incorporated many but for some I don't really see how this with improve the clarity of the readme which is supposed to be rather a quick piece of information.
