### 2025-11-04 rvandewater/recipys#26 @simonprovost
> [@simonprovost](https://github.com/simonprovost) I have added a diagram to show how recipies works [@simonprovost](https://github.com/simonprovost). I do see your other suggestions and I have incorporated many but for some I don't really see how this with improve the clarity of the readme which is supposed to be rather a quick piece of information.

Definitely agreed!  That's why I prefer this to be an open conversation about how to improve the clarity and effectiveness of your readme, which is completely great.  Could you please make a list of what has "changed" overall —— and a branch to which I could viz. the updated readme or if it's merge on your dev/main track just let me know of that.

By the way, if the readme is getting too long, perhaps the most practical option would be to include a tab in your documentation that explains Recipes in further detail, e.g. with a diagram, the backends/core concepts explanations etc. Have a thought, not requesting it but just sharing an extra option.

Note that I am more here to ensure that the best line of action has been taken, and the actual outcome of your OSS is your call of course

Let me know:)
