### 2025-11-04 rvandewater/recipys#29 @simonprovost
Brills @rvandewater ! Make sure to notify on the JOSS review that maybe the branch to look at is `main` and not `dev`. If both are the same version for the entire JOSS review then no need of course.

Cheers
