### 2025-11-05 7e57088
ruff
