### 2025-11-05 9cf760d
added pandas and polars example.
