### 2025-11-06 5c105eb
update role bug and massively improve docs and website
