+++ mkdocs.yml
@@ -9,8 +9,10 @@ nav:
   - Home: index.md
   - API Reference: api.md
   - Examples:
-      - Tutorial: examples/tutorial.ipynb
-      - Pandas and Polars: examples/pandas_or_polars.ipynb
+      - Basics: examples/tutorial.ipynb
+      - ICU data: examples/icu_advanced.ipynb
+      - Pandas or Polars: examples/pandas_or_polars.ipynb
+
 theme:
   name: material
   logo: figures/recipies_logo.svg
