### 2025-11-06 cabed1c
mkdocs workflow
