### 2025-11-06 rvandewater/recipys#26 @rvandewater
Hi @simonprovost
- I've added a diagram showing how ReciPies works
- Added uv documentation
- Incorporated most of your smaller suggestions
Let me know if you have any issues with the current state of the readme :).
