+++ mkdocs.yml
@@ -12,6 +12,7 @@ nav:
       - Basics: examples/tutorial.ipynb
       - ICU data: examples/icu_advanced.ipynb
       - Pandas or Polars: examples/pandas_or_polars.ipynb
+  - Concepts: concepts.md
 
 theme:
   name: material
