+++ .github/workflows/deploy-mkdocs.yaml
@@ -19,6 +19,8 @@ jobs:
       pages: write
       id-token: write
     steps:
+      - name: Checkout repository
+        uses: actions/checkout@v4
       - name: Setup Python
         uses: actions/setup-python@v5
         with:
@@ -34,5 +36,4 @@ jobs:
       - id: deployment
         uses: actions/deploy-pages@v4
         with:
-          # Path to your built site directory
           source: site
