### 2025-11-06 40c32d6
remove non-existent handler
