+++ mkdocs.yml
@@ -44,12 +44,6 @@ plugins:
       default_handler: python
       handlers:
         python:
-          rendering:
-            show_root_heading: true
-            show_source: false
-            members_order: source
-            merge_init_into_class: true
-            show_signature: true
   - mkdocs-jupyter:
       execute: true
       include_source: true
