### 2025-11-06 00cc44c
add recipies install to workflow
