### 2025-11-06 rvandewater/recipys#30 @rvandewater
Hi @simonprovost , thanks for the suggestions. I have overhauled the documentation. I've created an mkdocs website with better readable API documentation, which incorporates all of your comments. Some of the suggested auto-generated documents were kept because they provide a better overview of the internal workings of the package. Please check it out and see if you have any further requests!
