### 2025-11-06 #9261 @rvandewater
Hi @crvernon @simonprovost

Thanks for your work despite the large amount of work everyone has to do! I believe to I addressed a lot of minor and some major points including:
- An overhauled website with Jupyter examples that are run directly before compiling
- Updated a lot of minor typos and code quality issues
- Added a diagram to the readme and paper to show how ReciPies work in a schematic manner.
Please let me know about the next steps, in case you have any suggestions.
