### 2025-11-06 rvandewater/recipys#26 @simonprovost
Readme is brillant! At the very least, stable enough to grow over time, thanks for your promptness @rvandewater, very much appreciated!
