### 2025-11-06 rvandewater/recipys#30 @simonprovost
Brillant doc! At the very least, ready to grow over time with the development of Recipes! Thanks for your extra promptness on this @rvandewater, very much appreciated!
