+++ paper/paper.md
@@ -22,7 +22,7 @@ affiliations:
     name: Hasso Plattner Institute for Digital Health at Mount Sinai, Icahn School of Medicine at Mount Sinai, New York City, NY, USA
   - index: 3
     name: Innsbruck Medical University, Innsbruck, Austria
-date: 2025-07-25
+date: 2025-11-13
 bibliography: paper.bib
 repository: https://github.com/rvandewater/ReciPies
 tags:
@@ -30,114 +30,47 @@ tags:
 
 # Summary
 
-Machine Learning (ML) workflows live or die by their data‑preprocessing steps, yet in Python, these steps are often
-scattered across ad‑hoc scripts or opaque Scikit-Learn (sklearn) snippets that are hard to read, audit, or reuse.
-`ReciPies` provides a concise, human‑readable, and fully reproducible method to declare, execute, and share
-preprocessing pipelines, that adheres to Configuration as Code principles.
-It lets users describe transformations as a recipe made of ordered *steps* (e.g., imputing, encoding, normalizing)
-applied to variables identified by semantic roles (predictor, outcome, ID, time stamp, etc.). Recipes can be *prepped*
-once, *baked* many times, and separated between training and new data. `ReciPies` provides the choice of Pandas and
-Polars backends and is easily extensible. Data provenance can be tracked and published.
-Packaging preprocessing as clear, declarative objects, `ReciPies` lowers the cognitive load of feature engineering,
-improves reproducibility, and makes methodological choices explicit, benefiting individual researchers, engineering
-teams, and peer reviewers alike.
+Machine Learning (ML) workflows live or die by their data‑preprocessing steps. In Python, these steps are often
+scattered across ad‑hoc scripts or opaque scikit-learn snippets that are hard to read, audit, or reuse.
+`ReciPies` provides a concise, human‑readable, and reproducible way to declare, execute, and share
+preprocessing pipelines following configuration-as-code principles. It lowers the cognitive load of
+feature engineering, improves reproducibility, and makes methodological choices explicit for
+researchers, engineering teams, and peer reviewers.
 
 # Statement of need
 
-Robust machine‑learning results in science hinge on transparent, reproducible data‑preprocessing—yet in Python, these
-steps are typically spread across ad‑hoc notebooks or are buried inside opaque scripts; that is, if this code is even
-made available. Additionally, most variable semantics are unclear (called *roles*). These problems confound research
-results, complicate peer review and hinder reuse. Researchers and engineers working with longitudinal regulated data (
-e.g., energy production, finance, and environmental monitoring) especially need pipelines they are able to audit,
+Transparent and reproducible preprocessing remains a weak link in many scientific ML studies. The consequences are (1) confounded research results,
+(2) complicated peer review, and (3) poor reuse. Researchers and engineers working with longitudinal regulated data (e.g., in energy production, health, finance, or environmental monitoring) in particular need pipelines they can audit,
 serialize, and hand to collaborators without reverse‑engineering a tangle of imperative code [@10.1145/3641525]. The
-lack of reproducibility has been documented extensively in
+current lack of reproducibility has been documented extensively in the
 literature [@johnsonReproducibilityCriticalCare2017a; @kellyKeyChallengesDelivering2019a; @semmelrockReproducibilityMachinelearningbasedResearch2025].
 
-`ReciPies` fills this gap by bringing a tidy, stepwise *recipe* interface to Python. Users declare transformations over
-variables selected by semantic roles; recipes are "prepped" once on training data and "baked" on new data to eliminate
-leakage; and every step is inspectable, versionable, and serializable (JSON/YAML). `Recipes` runs on
[…]
