+++ paper/paper.bib
@@ -308,3 +308,22 @@
   organization = {Zenodo},
   file = {/Users/robin/Zotero/storage/Y2T2RU6L/15313097.html}
 }
+@article{reynaEarlyPredictionSepsis2020a,
+  title = {Early {{Prediction}} of {{Sepsis From Clinical Data}}: {{The PhysioNet}}/{{Computing}} in {{Cardiology Challenge}} 2019},
+  shorttitle = {Early {{Prediction}} of {{Sepsis From Clinical Data}}},
+  author = {Reyna, Matthew A. and Josef, Christopher S. and Jeter, Russell and Shashikumar, Supreeth P. and Westover, M. Brandon and Nemati, Shamim and Clifford, Gari D. and Sharma, Ashish},
+  date = {2020-02},
+  journaltitle = {Critical Care Medicine},
+  shortjournal = {Crit Care Med},
+  volume = {48},
+  number = {2},
+  eprint = {31939789},
+  eprinttype = {pubmed},
+  pages = {210--217},
+  issn = {1530-0293},
+  doi = {10.1097/CCM.0000000000004145},
+  langid = {english},
+  pmcid = {PMC6964870},
+  keywords = {Algorithms,Early Diagnosis,Electronic Health Records,Female,Humans,Intensive Care Units,Male,Sepsis,Severity of Illness Index,Time Factors,United States},
+  file = {/Users/robin/Zotero/storage/GDGXT3SK/Reyna et al. - 2020 - Early Prediction of Sepsis From Clinical Data The PhysioNetComputing in Cardiology Challenge 2019.pdf}
+}
