### 2025-11-13 0874532
update figures due to formatting issues
