+++ paper/paper.md
@@ -55,9 +55,9 @@ learning pipelines.
 # Design and implementation
 
 `ReciPies` adopts a tidy, stepwise *recipe* interface that emphasizes semantic roles over column names and a strict separation of fitting from application. Transformations are declared on roles such as predictor, outcome, identifier, or timestamp. Recipes are prepped on training data and baked on new data to prevent leakage. Each step is inspectable, versionable, and serializable to JSON or YAML for provenance and review. Steps are composable with explicit state and deterministic behavior given fixed inputs and seeds. `ReciPies` supports both Pandas [@mckinney-proc-scipy-2010], which is widely adopted in the ML community, and the more recent Polars [@vinkPolarsPolarsPython2024], which offers increased performance.
-![recipies flowchart](../docs/figures/recipies_flow.svg)
+![recipies flowchart](../docs/figures/recipies_flow.pdf)
 A typical workflow 1) loads a Pandas or Polars training `DataFrame`, 2) wraps it as an `Ingredients` object that records role metadata, 3) defines a `Recipe` from `Steps` operating on columns selected based on roles by `Selectors`, 4) preps the recipe on the training split to estimate parameters, and 5) bakes it on the held-out split to apply those parameters without leakage. The baked outputs feed downstream modeling and evaluation. Figure 1 gives an overview of this workflow.
-![physionet code snippet](../docs/figures/physionet_recipies.svg)
+![physionet code snippet](../docs/figures/recipies_code_snippet.pdf)
 Figure 2 demonstrates usage on the PhysioNet Computing in Cardiology 2019 dataset [@reynaEarlyPredictionSepsis2020a], including role assignment, time-aware imputation, categorical encoding, and normalization. The prepped recipe serializes to JSON or YAML, and reloading the artifact reproduces the transforms across supported platforms.
 
 Complete code and interactive notebooks are available in the project documentation. `ReciPies` also provides a benchmarking suite comparing the performance of different preprocessing steps
