### 2025-11-13 #9261 @rvandewater
I implemented fixes because the carbon SVG did not render on the first generated PDF, and the diagram rendered incorrectly.
