### 2025-11-17 #9261 @panagiotisanagnostou
Hello @crvernon, @simonprovost and @rvandewater ,

I have to apologize for the delayed response, but it is a really busy time for me.

That being said, I have several points that I believe warrant amendments in the repository in general.

I will respond with a new comment so I can add a checklist.
