### 2025-11-17 #9261 @panagiotisanagnostou
@rvandewater, these are the issues I have found so far. I want to conduct some more functional experiments of the package to see the results and if any further issues arise.

1. [ ] Fix the first and last references in the paper (correct casing and author names).
2. [ ] Generalize software execution to multiple platforms and Python versions; current CI workflows only use `ubuntu-latest` and `python-3.12`.
3. [ ] Remove the `.idea` directory from the repository and add it to `.gitignore`.
4. [ ] Increase test coverage; 85% is too low for this project.
5. [ ] Add automated code quality checks (e.g., linters / static analysis) instead of relying only on manual review.
6. [ ] Improve `benchmarking.ipynb`: add clearer comments, more descriptive markdown cells, and better-structured code cells so the workflow is easier to follow.
7. [ ] In the README installation section for developers (editable install), make it explicit that you are offering **two alternative ways** to install, and clearly label them as such.
8. [ ] Fix the documentation / paths that reference a `benchmarks` directory; it does not exist as currently described.
9. [ ] Extend the “basics” example with more details on how to actually execute the package (commands, expected outputs, typical workflow).
10. [ ] Adjust the printed table formatting (border and column sizes) so the tables are readable and consistently sized.
11. [ ] Investigate and fix the execution error in the **“ICU Preprocessing Pipeline with ReciPies”** example/pipeline.
12. [ ] Move or surface the **Concepts** section so it appears more prominently (e.g., near the Home page) – it contains very important information for users of the package.
13. [ ] Expand the benchmarking explanation: clearly state the units, the baseline, and the meaning of speed-up / speed-down, and justify the performance claims.
14. [ ] Add an example demonstrating how to **train a machine learning model** using the package end-to-end.
15. [ ] Generalize and align `requirements.txt` with the dependencies used in your workflows (avoid unnecessary constraints that conflict with CI).
16. [ ] Integrate Codecov and Codacy (or equivalent) with badges in the README, targeting at least **95% test coverage** and a **Codacy grade A**.

The order of issues as I discovered them while filling my checklist.
