### 2025-11-17 959959e
minor paper updates and bibliography issues
