+++ paper/paper.bib
@@ -299,7 +299,7 @@
 @software{warmerdamKoaningScikitlegoV0952025,
   title = {Koaning/Scikit-Lego: V0.9.5},
   shorttitle = {Koaning/Scikit-Lego},
-  author = {family=warmerdam, given=vincent, prefix=d, useprefix=false and Bruzzesi, Francesco and MBrouns and Collot, Stéphane and family=Boer, given=Josko, prefix=de, useprefix=false and Kübler, Robert and family=hoeven, prefix=pim-, useprefix=true and {mkalimeri} and Paulino, Arthur and Gorelli, Marco Edward and Verheijen, Peter and Borry, Maxime and Hoogland, Kay and Masip, David and Kowalczuk, Magdalena and {ktiamur} and AminaZ and Sharma, Gaurav and Lepelaars, Carlo and Rens and Cor and Hermans, Frits and Borrella, Tomas and Rose, Stephen Anthony and family=Dorsten, given=Sander, prefix=van, useprefix=false and Levitski, Hleb and Keromnes, Jan and Pérez-Lozao, Sergio Calderón and Payne, Skylar},
+  author = {family=Warmerdam, given=Vincent, prefix=D, useprefix=false and Bruzzesi, Francesco and MBrouns and Collot, Stéphane and family=Boer, given=Josko, prefix=de, useprefix=false and Kübler, Robert and family=hoeven, prefix=pim-, useprefix=true and {mkalimeri} and Paulino, Arthur and Gorelli, Marco Edward and Verheijen, Peter and Borry, Maxime and Hoogland, Kay and Masip, David and Kowalczuk, Magdalena and {ktiamur} and AminaZ and Sharma, Gaurav and Lepelaars, Carlo and Rens and Cor and Hermans, Frits and Borrella, Tomas and Rose, Stephen Anthony and family=Dorsten, given=Sander, prefix=van, useprefix=false and Levitski, Hleb and Keromnes, Jan and Pérez-Lozao, Sergio Calderón and Payne, Skylar},
   date = {2025-04-30},
   doi = {10.5281/zenodo.15313097},
   url = {https://zenodo.org/records/15313097},
