+++ requirements.txt
@@ -1,8 +1,4 @@
-pandas==2.3.0
-polars==1.31.0
-pyarrow==20.0.0
-pytest==8.4.1
-scikit-learn==1.7.0
-tqdm==4.67.1
-ruff==0.4.0
-memory_profiler==0.61.0
+pandas>=2.3.0
+polars[all]>=1.21.0
+scikit-learn>=1.5.2
+
