+++ pyproject.toml
@@ -31,7 +31,7 @@ dependencies = [
     "polars[all]>=1.21.0",
     "scikit-learn>=1.5.2"
 ]
-optional-dependencies = { dev = ["pytest>=7.0", "ruff>=0.1.0"], docs = [
+optional-dependencies = { dev = ["pytest>=7.0", "ruff>=0.1.0", "pytest-cov"], docs = [
     "mkdocs>=1.6",
     "mkdocs-material>=9.5",
     "mkdocstrings[python]>=0.26",
