+++ .github/workflows/code-coverage.yaml
@@ -17,7 +17,7 @@ jobs:
         uses: actions/setup-python@v4
 
       - name: Install dependencies
-        run: pip install -e ".[tests, polars, pandas]"
+        run: pip install -e ".[dev, polars, pandas]"
       - name: Run tests
         run: pytest --cov --cov-branch --cov-report=xml
 
