+++ environment.yml
@@ -1,9 +1,0 @@
-name: recipys_polars
-channels:
-  - conda-forge
-  - anaconda
-dependencies:
-  - python=3.12
-  - pip
-  - pip:
-      - -r requirements.txt
