### 2025-11-18 e1c3792
extra tests for recipe
