### 2025-11-18 fb18ff7
Added a lot of tests
