+++ requirements.txt
@@ -1,4 +1,3 @@
 pandas>=2.3.0
 polars[all]>=1.21.0
 scikit-learn>=1.5.2
-
