### 2025-11-18 97a11ef
extra edge cases
