+++ tests/test_ingredients.py
@@ -196,3 +196,205 @@ def test_ingredients_roles_copy():
     ingr_copy = Ingredients(ingr)
     assert ingr_copy.roles == roles
     assert ingr_copy.data.equals(ingr.data)
+
+
+def test_ingredients_init_invalid_roles(example_pl_df):
+    # Test roles with invalid column names
+    with pytest.raises(ValueError, match="Roles contains variable names that are not in the data"):
+        Ingredients(example_pl_df, roles={"invalid_column": ["role"]})
+
+
+def test_ingredients_init_backend_inference(example_pl_df):
+    # Test backend inference for polars
+    ingredients = Ingredients(example_pl_df)
+    assert ingredients.get_backend() == Backend.POLARS
+
+    # Test backend inference for pandas
+    pandas_df = example_pl_df.to_pandas()
+    ingredients = Ingredients(pandas_df)
+    assert ingredients.get_backend() == Backend.PANDAS
+
+
+def test_ingredients_add_role(example_pl_df):
+    # Test adding a role to a column with existing roles
+    ingredients = Ingredients(example_pl_df, roles={"x1": ["predictor"]})
+    ingredients.add_role("x1", "new_role")
+    assert "new_role" in ingredients.roles["x1"]
+
+    # Test adding a role to a column without roles (should raise an error)
+    with pytest.raises(RuntimeError, match="has no roles yet, use update_role instead"):
+        ingredients.add_role("x2", "new_role")
+
+
+def test_ingredients_update_role(example_pl_df):
+    # Test updating a role for a column
+    ingredients = Ingredients(example_pl_df, roles={"x1": ["predictor"]})
+    ingredients.update_role("x1", "new_predictor", "predictor")
+    assert "new_predictor" in ingredients.roles["x1"]
+    assert "predictor" not in ingredients.roles["x1"]
+
+    # Test updating a role without specifying old_role
+    ingredients.update_role("x1", "another_role")
+    assert ingredients.roles["x1"] == ["another_role"]
+
+    # Test invalid old_role
+    with pytest.raises(
+        ValueError, match="Attempted to set role of x1 from invalid_role to new_role but invalid_role not among current roles"
+    ):
+        ingredients.update_role("x1", "new_role", "invalid_role")
+
+
+def test_ingredients_select_dtypes(example_pl_df):
+    # Test selecting columns by data types
+    ingredients = Ingredients(example_pl_df)
+    selected = ingredients.select_dtypes(include=["float64"] if ingredients.get_backend() == Backend.PANDAS else ["Float64"])
+    assert "x1" in selected or "y" in selected
+
+
+def test_ingredients_get_dtypes(example_pl_df):
+    # Test retrieving data types
+    ingredients = Ingredients(example_pl_df)
+    dtypes = ingredients.get_dtypes()
+    assert len(dtypes) == len(example_pl_df.columns)
+
+
+def test_ingredients_get_str_dtypes(example_pl_df):
+    # Test retrieving data types as strings
+    ingredients = Ingredients(example_pl_df)
+    str_dtypes = ingredients.get_str_dtypes()
+    assert all(isinstance(dtype, str) for dtype in str_dtypes.values())
+
+
+def test_ingredients_groupby(example_pl_df):
+    # Test grouping by columns for polars backend
+    ingredients = Ingredients(example_pl_df)
+    grouped = ingredients.groupby("x1")
+    assert grouped is not None
+
+    # Test grouping by columns for pandas backend
+    pandas_df = example_pl_df.to_pandas()
+    ingredients = Ingredients(pandas_df)
+    grouped = ingredients.groupby("x1")
+    assert grouped is not None
+
+
+def test_ingredients_getitem(example_pl_df):
[…]
