### 2025-11-19 1205f57
Add Codacy badge to README
Added Codacy badge for code quality monitoring.
