+++ .github/workflows/ci.yml
@@ -13,6 +13,9 @@ on:
 jobs:
   lint:
     runs-on: ubuntu-latest
+    strategy:
+      matrix:
+        python-version: ["3.9", "3.10", "3.11", "3.12"]
 
     steps:
       - name: Checkout code
@@ -21,7 +24,7 @@ jobs:
       - name: Set up Python
         uses: actions/setup-python@v4
         with:
-          python-version: "3.12.0"
+          python-version: ${{ matrix.python-version }}
 
       - name: Install dependencies
         run: |
