### 2025-11-19 a397696
more version details and remove benchmarking notebook
