+++ README.md
@@ -57,13 +57,16 @@ uv add recipies
 
 ```bash
 # with conda (optional)
-conda env update -f environment.yml
+conda env create -n ReciPies python=3.12
 conda activate ReciPies
-# with pip
+```
+Then, from the root of the repository, run with pip:
+```bash
 pip install -e .
-# with uv venv
+```
+Or with uv (if you have uv installed):
+``` bash
 uv venv && source .venv/bin/activate
-
 ```
 
 ## Getting Start
