### 2025-11-19 1fe3b43
Additions for benchmarking, added versions. Clarified readme.
