+++ mkdocs.yml
@@ -13,7 +13,7 @@ nav:
       - Basics: examples/tutorial.ipynb
       - ICU data: examples/icu_advanced.ipynb
       - Pandas or Polars: examples/pandas_or_polars.ipynb
-
+  - Benchmarking: benchmarking.md
 theme:
   name: material
   logo: figures/recipies_logo.svg
