### 2025-11-19 643d77c
remove python requires because it fails the build workflow
