### 2025-11-19 1e7de84
Hotfix python3.9: Add future annotations import to ingredients.py
