+++ src/recipies/ingredients.py
@@ -3,7 +3,7 @@ import pandas as pd
 import polars as pl
 from typing import overload
 from .constants import Backend
-
+from __future__ import annotations
 
 class Ingredients:
     """Wrapper around either polars.DataFrame to store columns roles (e.g., predictor)
