+++ docs/benchmarking.md
@@ -2,17 +2,18 @@
 
 This page provides an overview of the benchmarking process implemented in the `perform_benchmark.py` script. The benchmarking evaluates the performance of various preprocessing steps in the `ReciPies` library using different data sizes and backends (Polars and Pandas).
 
----
+______________________________________________________________________
 
 ## Overview
 
 The benchmarking script measures the **execution time** and **memory usage** of preprocessing steps applied to synthetic ICU data. The results are aggregated and saved as a CSV file for further analysis.
 
----
+______________________________________________________________________
 
 ## How It Works
 
 ### 1. **Synthetic Data Generation**
+
 The script uses the `generate_icu_data` function to create synthetic ICU data with configurable:
 
 - **Data sizes**: Number of rows in the dataset.
@@ -20,6 +21,7 @@ The script uses the `generate_icu_data` function to create synthetic ICU data wi
 - **Random seeds**: For reproducibility.
 
 ### 2. **Backends**
+
 The benchmarking supports two backends:
 
 - **Polars**: A high-performance DataFrame library.
@@ -28,29 +30,32 @@ The benchmarking supports two backends:
 The script converts the synthetic data to the appropriate backend format before running the benchmarks.
 
 ### 3. **Preprocessing Steps**
+
 The following preprocessing steps are benchmarked:
 
 - **Imputation**:
-  - `StepImputeFill`: Forward-fill and zero-fill strategies.
-  - `StepSklearn`: Using `MissingIndicator` from `sklearn`.
+    - `StepImputeFill`: Forward-fill and zero-fill strategies.
+    - `StepSklearn`: Using `MissingIndicator` from `sklearn`.
 - **Scaling**:
-  - `StepSklearn`: Using scalers like `StandardScaler`, `MinMaxScaler`, etc.
+    - `StepSklearn`: Using scalers like `StandardScaler`, `MinMaxScaler`, etc.
 - **Discretization**:
-  - `StepSklearn`: Using `KBinsDiscretizer`.
+    - `StepSklearn`: Using `KBinsDiscretizer`.
 - **Historical Accumulation**:
-  - `StepHistorical`: Using accumulators like `MIN`, `MAX`, `MEAN`, and `COUNT`.
+    - `StepHistorical`: Using accumulators like `MIN`, `MAX`, `MEAN`, and `COUNT`.
 
 ### 4. **Metrics**
+
 For each preprocessing step, the script measures:
 
 - **Execution Time**: The time taken to apply the step.
 - **Memory Usage**: The peak memory usage during the step.
 
----
+______________________________________________________________________
 
 ## Running the Benchmark
 
 ### Command
+
 Run the script using the following command:
 
 ```bash
@@ -58,19 +63,22 @@ python perform_benchmark.py --data_sizes 1000 10000 --seeds 42 41
 ```
 
 ### Arguments
+
 - `--data_sizes`: A list of data sizes to benchmark (e.g., `1000`, `10000`).
 - `--seeds`: A list of random seeds for reproducibility.
 
 ### Example
+
 ```bash
 python perform_benchmark.py --data_sizes 1000 10000 100000 --seeds 42 43
 ```
 
----
+______________________________________________________________________
 
 ## Results
 
 ### Output
+
 The script generates a CSV file with the benchmarking results. The filename includes the data sizes, seeds, and a timestamp. For example:
 
 ```
@@ -78,6 +86,7 @@ results_datasizes_[1000, 10000]_seeds_[42, 43]_datetime_2025-11-19_12-00-00.csv
 ```
 
 ### Aggregated Metrics
+
 The results include:
 
 - **Mean Execution Time** (`duration_mean`)
@@ -87,48 +96,54 @@ The results include:
[…]
