### 2025-11-19 #9261 @rvandewater
Hi @panagiotisanagnostou:

 _1. [ ]  Fix the first and last references in the paper (correct casing and author names)._
 I found at least one issue with the author name and fixed it. I am not sure if there was another issue you found?
_2. [ ]  Generalize software execution to multiple platforms and Python versions; current CI workflows only use `ubuntu-latest` and `python-3.12`._
I now use python 3.9-3.13. What OS's would you suggest we use as here we can have exponentially many combinations?
_3. [ ]  Remove the `.idea` directory from the repository and add it to `.gitignore`._
Done
_4. [ ]  Increase test coverage; 85% is too low for this project._
I have added a lot of tests and coverage integration with a badge.
_5. [ ] Add automated code quality checks (e.g., linters / static analysis) instead of relying only on manual review._
There is already automated pre-commit check that runs in a workflow. What do you suggest we add?
_6. [ ] Improve `benchmarking.ipynb`: add clearer comments, more descriptive markdown cells, and better-structured code cells so the workflow is easier to follow._
This file was a leftover and has been removed.
_7. [ ] In the README installation section for developers (editable install), make it explicit that you are offering **two alternative ways** to install, and clearly label them as such._
Done.
_8. [ ] Fix the documentation / paths that reference a `benchmarks` directory; it does not exist as currently described._
I believe I removed all references now.
_9. [ ]  Extend the “basics” example with more details on how to actually execute the package (commands, expected outputs, typical workflow)._
I am not sure what you would like us to add as there is a typical workflow in that page. To be sure, you are looking at this page, right: https://rvandewater.github.io/ReciPies/examples/tutorial/.
_10. [ ]  Adjust the printed table formatting (border and column sizes) so the tables are readable and consistently sized._
Where do you mean exactly? The tables in the notebooks are quite readable for me. Maybe you have a screenshot to show what you mean?
_11. [ ]  Investigate and fix the execution error in the **“ICU Preprocessing Pipeline with ReciPies”** example/pipeline._
Fixed.
_12. [ ]   Move or surface the **Concepts** section so it appears more prominently (e.g., near the Home page) – it contains very important information for users of the package._
Concepts has been moved next to the home page as a tab.
_13. [ ]   Expand the benchmarking explanation: clearly state the units, the baseline, and the meaning of speed-up / speed-down, and justify the performance claims._
Added a documentation page that explains the benchmarking in detail (https://rvandewater.github.io/ReciPies/benchmarking/). Let me know if you're missing something
_14. [ ]   Add an example demonstrating how to **train a machine learning model** using the package end-to-end._
This is already an end-to-end example in this example: https://rvandewater.github.io/ReciPies/examples/icu_advanced/#5-train-a-machine-learning-model. I'm not sure if you had anything else in mind?
_15. [ ]   Generalize and align `requirements.txt` with the dependencies used in your workflows (avoid unnecessary constraints that conflict with CI)._
Done.
_16. [ ]   Integrate Codecov and Codacy (or equivalent) with badges in the README, targeting at least **95% test coverage** and a **Codacy grade A**._
Done, added almost 100 extra tests. There is now a 97% coverage and we have a codeacy grade A.
