### 2025-11-19 7e2d747
update test to work with 3.9
