+++ .github/workflows/deploy-mkdocs.yaml
@@ -24,7 +24,7 @@ jobs:
       - name: Setup Python
         uses: actions/setup-python@v5
         with:
-          python-version: "3.12"
+          python-version: "3.13"
       - name: Install dependencies
         run: pip install mkdocs mkdocs-material mkdocstrings[python] mkdocs-jupyter recipies
       - name: Build MkDocs site
