### 2025-11-24 ea17794
automated run of tests every two weeks
