### 2025-11-24 cf191c3
pre-commit and empty notebooks
