+++ .github/workflows/tests.yaml
@@ -6,7 +6,7 @@ on:
   pull_request:
     branches: [main, "release/*", "dev"]
   schedule:
-    - cron: "0 0 */14 * *"  # Runs every two weeks at midnight UTC
+    - cron: "0 0 */14 * *" # Runs every two weeks at midnight UTC
 
 jobs:
   run_tests_ubuntu:
