### 2025-11-24 2fcfd64
adjust to mkdocs
