### 2025-11-24 c60de24
change mkdocs to recipies and reformat ipynb
