### 2025-11-24 989057b
recipe display without borders.
