+++ src/recipies/recipe.py
@@ -184,7 +184,9 @@ class Recipe:
 
         # Print all existing roles and how many variables are assigned to each
         num_roles = Counter(chain.from_iterable(self.data.roles.values()))
-        num_roles = pl.DataFrame({"role": [r for r in num_roles.keys()], "#variables": [n for n in num_roles.values()]})
+        num_roles = pd.DataFrame(
+            {"role": [r for r in num_roles.keys()], "amount of variables": [n for n in num_roles.values()]}
+        )
         repr += "Inputs:\n\n" + num_roles.__repr__() + "\n\n"
 
         # Print all steps
