+++ paper/paper.bib
@@ -29,6 +29,20 @@
   file = {/Users/robin/Zotero/storage/V3LYVWF4/Galli - 2021 - Feature-engine A Python package for feature engineering for machine learning.pdf}
 }
 
+@inproceedings{gundersenStateArtReproducibility2018,
+  title = {State of the Art: Reproducibility in Artificial Intelligence},
+  shorttitle = {State of the Art},
+  booktitle = {Proceedings of the {{Thirty-Second AAAI Conference}} on {{Artificial Intelligence}} and {{Thirtieth Innovative Applications}} of {{Artificial Intelligence Conference}} and {{Eighth AAAI Symposium}} on {{Educational Advances}} in {{Artificial Intelligence}}},
+  author = {Gundersen, Odd Erik and Kjensmo, Sigbjørn},
+  date = {2018-02-02},
+  series = {{{AAAI}}'18/{{IAAI}}'18/{{EAAI}}'18},
+  pages = {1644--1651},
+  publisher = {AAAI Press},
+  location = {New Orleans, Louisiana, USA},
+  abstract = {Background: Research results in artificial intelligence (AI) are criticized for not being reproducible. Objective: To quantify the state of reproducibility of empirical AI research using six reproducibility metrics measuring three different degrees of reproducibility. Hypotheses: 1) AI research is not documented well enough to reproduce the reported results. 2) Documentation practices have improved over time. Method: The literature is reviewed and a set of variables that should be documented to enable reproducibility are grouped into three factors: Experiment, Data and Method. The metrics describe how well the factors have been documented for a paper. A total of 400 research papers from the conference series IJCAI and AAAI have been surveyed using the metrics. Findings: None of the papers document all of the variables. The metrics show that between 20\% and 30\% of the variables for each factor are documented. One of the metrics show statistically significant increase over time while the others show no change. Interpretation: The reproducibility scores decrease with increased documentation requirements. Improvement over time is found. Conclusion: Both hypotheses are supported.},
+  isbn = {978-1-57735-800-8}
+}
+
 @inproceedings{j.PyjanitorCleanerAPI2019,
   title = {Pyjanitor: {{A Cleaner API}} for {{Cleaning Data}}},
   shorttitle = {Pyjanitor},
@@ -80,7 +94,6 @@
[…]
