### 2026-01-02 #9261 @rvandewater
Hi @crvernon thanks and happy new year!

- I have addressed the formatting issues/spacing in the paper.
- I have clarified the claims with specific papers. The proceedings as a whole are referenced in a separate sentence, as the main point of citing this was to show the development of this issue within the ML field. Please let me know if this acceptable or you prefer a different solution.
- I have created a zenodo release according to your suggestion: [10.5281/zenodo.15863617](https://doi.org/10.5281/zenodo.15863617).

Thanks!
