### 2026-01-02 #9261 @crvernon
@rvandewater see the following error for the missing citation:

> ⚠️ Couldn't acccept/publish paper. [An error happened](https://github.com/openjournals/joss-papers/actions/runs/20662516160?check_suite_focus=true). The generated XML metadata file is invalid.
>
> IDREFS attribute rid references an unknown ID "ref-reynaEarlyPredictionSepsis2020a"
