### 2026-01-03 5c03509
update missing dois and missing references.
