+++ paper/paper.bib
@@ -90,7 +90,7 @@
   pages = {361--376},
   publisher = {PMLR},
   issn = {2640-3498},
-  url = {https://proceedings.mlr.press/v68/johnson17a.html},
+  url = {https://doi.org/10.1101/2024.06.04.24308417},
   urldate = {2025-07-23},
   abstract = {Mortality prediction of intensive care unit (ICU) patients facilitates hospital benchmarking and has the opportunity to provide caregivers with useful summaries of patient health at the bedside. The development of novel models for mortality prediction is a popular task in machine learning, with researchers typically seeking to maximize measures such as the area under the receiver operator characteristic curve (AUROC). The number of ’researcher degrees of freedom’ that contribute to the performance of a model, however, presents a challenge when seeking to compare reported performance of such models. In this study, we review publications that have reported performance of mortality prediction models based on the Medical Information Mart for Intensive Care (MIMIC) database and attempt to reproduce the cohorts used in their studies. We then compare the performance reported in the studies against gradient boosting and logistic regression models using a simple set of features extracted from MIMIC. We demonstrate the large heterogeneity in studies that purport to conduct the single task of ’mortality prediction’, highlighting the need for improvements in the way that prediction tasks are reported to enable fairer comparison between models. We reproduced datasets for 38 experiments corresponding to 28 published studies using MIMIC. In half of the experiments, the sample size we acquired was 25\% greater or smaller than the sample size reported. The highest discrepancy was 11,767 patients. While accurate reproduction of each study cannot be guaranteed, we believe that these results highlight the need for more consistent reporting of model design and methodology to allow performance improvements to be compared. We discuss the challenges in reproducing the cohorts used in the studies, highlighting the importance of clearly reported methods (e.g. data cleansing, variable selection, cohort selection) and the need for open code and publicly available benchmarks.},
   eventtitle = {Machine {{Learning}} for {{Healthcare Conference}}},
@@ -121,7 +121,7 @@
   shorttitle = {Recipes},
   author = {Kuhn, Max and Wickham, Hadley and Hvitfeldt, Emil and Software, Posit and PBC},
   date = {2024-07-04},
-  url = {https://cloud.r-project.org/web/packages/recipes/index.html},
+  url = {https://doi.org/10.32614/CRAN.package.recipes},
   urldate = {2024-07-09},
   abstract = {A recipe prepares your data for modeling. We provide an extensible framework for pipeable sequences of feature engineering steps provides preprocessing tools to be applied to data. Statistical parameters for the steps can be estimated from an initial data set and then applied to other data sets. The resulting processed output can then be used as inputs for statistical or machine learning models.},
   version = {1.1.0},
@@ -163,8 +163,7 @@
   number = {85},
   pages = {2825--2830},
   issn = {1533-7928},
-  doi = {10.5555/1953048.2078195},
-  url = {http://jmlr.org/papers/v12/pedregosa11a.html},
+  url = {https://dl.acm.org/doi/10.5555/1953048.2078195},
   urldate = {2022-12-21},
   abstract = {Scikit-learn is a Python module integrating a wide range of state-of-the-art machine learning algorithms for medium-scale supervised and unsupervised problems. This package focuses on bringing machine learning to non-specialists using a general-purpose high-level language. Emphasis is put on ease of use, performance, documentation, and API consistency. It has minimal dependencies and is distributed under the simplified BSD license, encouraging its use in both academic and commercial settings. Source code, binaries, and documentation can be downloaded from http://scikit-learn.sourceforge.net.}
 }
@@ -191,7 +190,8 @@
   pages = {5485--5495},
   publisher = {Curran Associates Inc.},
   location = {Red Hook, NY, USA},
-  doi = {10.5555/3454287.3454779},
[…]
